Agrego lógica para mostrar figuritas de tipo estadio en el album de un usuario

+++ TPCuatrimestral-Album/dominio2/Figurita.cs
@@ -13,7 +13,7 @@ namespace dominio2
         public TipoDeFigurita TipoDeFigurita { get; set; }
         //public int IDAlbum { get; set; }
         // public int PrecioVenta { get; set; }
-        public int Ubicacion { get; set; }
+        public Int16 Ubicacion { get; set; }
         public bool Pegada { get; set; }
         public bool Activo { get; set; }
         //public Jugador Jugador { get; set; }
